moved localStorage call to client comp
